# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "My API",
        "description": "REST API for the sonotrak application."
    },
...
    "host": "localhost:8080",
    "basePath": "/",
    "schemes": [
        "http"
    ],
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "definitions": {
...
        "route": {
            "type": "object",
            "properties": {
                "agency_id": {
                    "type": "string"
                },
                "route_short_name": {
                    "type": "string"
                },
                "route_long_name": {
                    "type": "string"
                },
                "route_desc": {
                    "type": "string"
                },
                "route_type": {
                    "type": "integer"
                },
                "route_url": {
                    "type": "string"
                },
                "route_color": {
                    "type": "string"
                },
                "route_text_color": {
                    "type": "string"
                }
            }
        },
...
        "stop": {
            "type": "object",
            "properties": {
                "stop_code": {
                    "type": "string"
                },
                "stop_name": {
                    "type": "string"
                },
                "stop_desc": {
                    "type": "string"
                },
                "stop_lon": {
                    "type": "string"
                },
                "stop_lat": {
                    "type": "string"
                }
            }
        },
        "stoptime": {
            "type": "object",
            "properties": {
                "trip_id": {
                    "type": "string"
                },
                "stop_id": {
                    "type": "string"
                },
                "arrival_time": {
                    "type": "string"
                },
                "departure_time": {
                    "type": "string"
                },
                "stop_headsign": {
                    "type": "string"
                },
                "stop_sequence": {
                    "type": "integer"
                },
                "pickup_type": {
                    "type": "integer"
                }
            }
        },
...
    },
    "paths": {
...
        "/api/route/{id}": {
            "get": {
                "summary": "get route by id ",
                "description": "Updates an existing route with new information",
                "tags": [
                    "Route"
                ],
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/route"
                        }
                    },
                    "401": {
                        "description": "Unauthenticated!"
                    }
                }
            },
            "put": {
                "summary": "Update an existing route",
                "description": "Updates an existing route with new information",
                "tags": [
                    "Route"
                ],
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "user",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/route"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/route"
                        }
                    },
                    "401": {
                        "description": "Unauthenticated!"
                    }
                }
            },
            "delete": {
                "summary": "delete route by id ",
                "description": "delete an existing route",
                "tags": [
                    "Route"
                ],
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/route"
                        }
                    },
                    "401": {
                        "description": "Unauthenticated!"
                    }
                }
            }
        },
...
        "/api/stoptime/{id}": {
            "get": {
                "summary": "get stoptime by id ",
                "description": "Updates an existing stoptime with new information",
                "tags": [
                    "Stoptimes"
                ],
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/stoptime"
                        }
                    },
                    "401": {
                        "description": "Unauthenticated!"
                    }
                }
            },
            "put": {
                "summary": "Update an existing stoptime",
                "description": "Updates an existing stoptime with new information",
                "tags": [
                    "Stoptimes"
                ],
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "user",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/stoptime"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/stoptime"
                        }
                    },
                    "401": {
                        "description": "Unauthenticated!"
                    }
                }
            },
            "delete": {
                "summary": "delete stoptime by id ",
                "description": "delete an existing stoptime",
                "tags": [
                    "Stoptimes"
                ],
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/stoptime"
                        }
                    },
                    "401": {
                        "description": "Unauthenticated!"
                    }
                }
            }
        },
...
        "/api/stop/{id}": {
            "get": {
                "summary": "get stop by id ",
                "description": "Updates an existing stop with new information",
                "tags": [
                    "Stop"
                ],
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/stop"
                        }
                    },
                    "401": {
                        "description": "Unauthenticated!"
                    }
                }
            },
            "put": {
                "summary": "Update an existing stop",
                "description": "Updates an existing stop with new information",
                "tags": [
                    "Stop"
                ],
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "user",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/stop"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/stop"
                        }
                    },
                    "401": {
                        "description": "Unauthenticated!"
                    }
                }
            },
            "delete": {
                "summary": "delete stop by id ",
                "description": "delete an existing stop",
                "tags": [
                    "Stop"
                ],
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/stop"
                        }
                    },
                    "401": {
                        "description": "Unauthenticated!"
                    }
                }
            }
        },
...
    }
}